Update: change main logic
The previous version had a big problem. There was no logic problem but the website had a problem.
Sometimes they change the post number. Because they add a post in middle of posts and delete a post. It makes the number shifted
So I changed the logic.
This no longer identifies by a post number whether a new post has been posted.